[Change] Continu on the login server work

+++ general/base/ProtocolParsingCheck.h
@@ -18,22 +18,22 @@ class ProtocolParsingCheck : public ProtocolParsingBase
         bool valid;
         bool parseIncommingDataRaw(const char *commonBuffer, const quint32 &size,quint32 &cursor);
     private:
-        void parseMessage(const quint8 &mainCodeType,const char *data,const int &size);
-        void parseFullMessage(const quint8 &mainCodeType,const quint8 &subCodeType,const char *data,const int &size);
+        void parseMessage(const quint8 &mainCodeType,const char *data,const unsigned int &size);
+        void parseFullMessage(const quint8 &mainCodeType,const quint8 &subCodeType,const char *data,const unsigned int &size);
         //have query with reply
-        void parseQuery(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const int &size);
-        void parseFullQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
+        void parseQuery(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const unsigned int &size);
+        void parseFullQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const unsigned int &size);
         //send reply
-        void parseReplyData(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const int &size);
-        void parseFullReplyData(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
+        void parseReplyData(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const unsigned int &size);
+        void parseFullReplyData(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const unsigned int &size);
         //message
         void errorParsingLayer(const QString &error);
         void messageParsingLayer(const QString &message) const;
 
         void disconnectClient();
 
-        ssize_t read(char * data, const int &size);
-        ssize_t write(const char * data, const int &size);
+        ssize_t read(char * data, const size_t &size);
+        ssize_t write(const char * data, const size_t &size);
 };
 
 }
